requirements.txt
numpy
pillow
tensorflow
fastapi
uvicorn
python-multipart
streamlit
requests
pandas
matplotlib
